// types/meteora.ts
export interface MeteoraDAMMV2PoolData {
  alpha_vault: string;
  apr: number;
  base_fee: number;
  collect_fee_mode: number;
  created_at_slot: number;
  created_at_slot_timestamp: number;
  creator: string;
  dynamic_fee: number;
  farm_active: boolean;
  fee24h: number;
  fee_scheduler_mode: number;
  fee_tvl_ratio: number;
  has_farm: boolean;
  launchpad: string;
  liquidity: string;
  max_price: string;
  min_price: string;
  permanent_lock_liquidity: string;
  pool_address: string;
  pool_name: string;
  pool_price: number;
  pool_type: number;
  sqrt_max_price: string;
  sqrt_min_price: string;
  sqrt_price: number;
  token_a_amount: number;
  token_a_amount_usd: number;
  token_a_mint: string;
  token_a_symbol: string;
  token_a_vault: string;
  token_b_amount: number;
  token_b_amount_usd: number;
  token_b_mint: string;
  token_b_symbol: string;
  token_b_vault: string;
  tokens_verified: boolean;
  tvl: number;
  updated_at: number;
  virtual_price: number;
  volume24h: number;
}